trying proxy.ts
